js edit 1 - missing html rep
