+++ HTML/ireland.html
@@ -1,27 +1,44 @@
 <!doctype html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <title>Automated World</title>
-        <link rel="stylesheet" href="./styles/main.css">
-    </head>
 
-    <body>
-        <header>
-            <nav>
-                <ul>
-                    <li><a href="../index.html">Home</a></li>
-                    <li><a href="./germany.html">Germany</a></li>
-                    <li><a class="active" href="./ireland.html">Ireland</a></li>
-                    <li><a href="./italy.html">Italy</a></li>
-                    <li><a href="./newzealand.html">New Zealand</a></li>
-                  </ul>
-            </nav>
-        </header>
-        <main id="container">
+<head>
+    <meta charset="utf-8">
+    <title>Automated World</title>
+    <link rel="stylesheet" href="../styles/main.css">
+</head>
 
-        </main>
+<body>
+    <header>
+        <h1>Travel Agency</h1>
+        <nav>
+            <ul>
+                <li><a href="../index.html">Home</a></li>
+                <li><a href="./germany.html">Germany</a></li>
+                <li><a class="active" href="./ireland.html">Ireland</a></li>
+                <li><a href="./italy.html">Italy</a></li>
+                <li><a href="./newzealand.html">New Zealand</a></li>
+            </ul>
+        </nav>
+    </header>
 
-        <script type="module" src="./scripts/main.js"></script>
-    </body>
+    <main id="container">
+        <div class="irelandImage">
+            <img src="../images/Ireland.jpg" alt="Ireland Countryside">
+        </div>
+        <h2>Ireland</h2>
+        <div class="countryInfo">
+            <div class="cities">
+                
+            </div>
+            <div class="landmarks">
+                <h3>Landmarks</h3>
+            </div>
+            <div class="citizens"></div>
+        </div>
+
+    </main>
+    <footer>&copy; Nutty Neighbors 2021</footer>
+    <script type="module" src="./scripts/main.js"></script>
+</body>
+
 </html>
