+++ HTML/italy.html
@@ -10,7 +10,7 @@
         <main id="container">
 
         </main>
-        <footer>&copy; Super Saturnians 2021</footer>
+        <footer>&copy; Nutty Neighbors 2021</footer>
         <script type="module" src="./scripts/main.js"></script>
     </body>
 </html>
