--- scripts/ireland/irelandMain.js
@@ -1,54 +1,0 @@
-//  1. create objects for cities landmarks and citizens & store them in a variable
-//  2. create a function to retrieve our objects
-//  3. 
-
-const cities = [
-    {
-        name: "Dublin",
-        population: "1,228,000",
-        funFact: "Dublin is home to the widest street in Europe. O’Connell Street measures 49m in width."
-    },
-    {
-        name: "Galway",
-        population: "79,934",
-        funFact: "Home to the Longest Place Name in Ireland. There is a city by the name of Muckanaghederdauhaulia"
-    },
-    {
-        name: "Killarney",
-        population: "14,504",
-        funFact: "Killarney means 'Church of the Blackthorn,’ and nobody is quite sure why."
-    }
-];
-
-const landmarks = [
-    {
-        name: "Rock of Cashel",
-        location: "Cashel, County Tipperary, Ireland"
-    },
-    {
-        name: "Newgrange Tomb",
-        location: "Boyne Valley, County Meath, Ireland"
-    },
-    {
-        name: "Blarney Stone and Castle",
-        location: "Blarney, County Cork, Ireland"
-    }
-];
-
-const famousCitizens = [
-    {
-        name: "Liam Neeson",
-        occupation: "Actor",
-        age: 68
-    },
-    {
-        name: "ConorMcGregor",
-        occupation: "Fighter",
-        age: 32
-    },
-    {
-        name: "Graham Norton",
-        occupation: "Talkshow Host",
-        age: 57
-    }
-];
