+++ styles/ireland.css
@@ -1,0 +1,3 @@
+.irelandImage {
+    width:auto
+}
